airflow/requirements-airflow.txt
# Python packages required for Airflow DAGs
# Install these in the Airflow container

psycopg2-binary==2.9.9
pandas==2.2.0
pyarrow==15.0.0
clickhouse-driver==0.2.7
python-dotenv==1.0.1
